requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.12.1
weasyprint==62.1
reportlab==4.2.0
structlog==23.2.0
slowapi==0.1.9
backports.zoneinfo==0.2.1; python_version<"3.9"